Download tracker data to buffer instead of writing to file
Writing tracker data to a file is dangerous: the http client is
async and may overwrite the data received from a tracker with
data of another download, before it is processed.

This commit writes data to a buffer instead of writing data to
a file and then loading it into memory.